Added first version of the player gun
Player can now walk up to a "console" and enter the gun.
When the player enters the gun, controls switch to gun controls and the camera switches to gun camera.
Player can look around in the gun (need to find a way to make it not disorienting).
Player can Exit the gun by hitting the same buttons used to enter it ('E' or 'F' work for now).